Move D8/DX lens related props into per-platform files
