+++ src/lens.h
@@ -1,5 +1,10 @@
 #ifndef _lens_h_
 #define _lens_h_
+
+#if defined(CONFIG_DIGIC_VIII) || defined(CONFIG_DIGIC_X)
+#include <platform/lens.h>
+#endif
+
 /** \file
  * Lens and camera control
  *
@@ -206,333 +211,6 @@ struct prop_lv_lens
 SIZE_CHECK_STRUCT( prop_lv_lens, 58 );
 
 #endif
-
-#if defined(CONFIG_DIGIC_8X)
-/* Digic 8 brings new properties:
- * PROP_LENS_STATIC_DATA  = PROP_LENS + PROP_LENS_NAME + ???
- * PROP_LENS_DYNAMIC_DATA = PROP_LV_LENS + ???
- *
- * Those are quite huge, they size depend on camera.
- * So far we have data from following models:
- *             M50   SX740    R     RP    SX70   250D   850D   R5/R6
- * STATIC     0x138  0x178  0x184  0x184  0x180  0x180  0x1C8  0x1C8
- * DYNAMIC    0x84   0x8C   0x90   0x90   0x90   0x8C   0x90   0x94
- * DryOS ICU   P2     P3     P4     P4     P4     P5     P8     P9
- *
- * A lot of PROP_LENS_STATIC_DATA can be decoded via `readid` evshell function.
- * Some can be decoded via ShootingInfoEx (models like SX70 don't have readid...)
- *
- * Structs seems to have `packed` attribute set, thus fields moving left and
- * right between models. For easier debugging I left those paddings filled in.
- */
-
-#if defined(CONFIG_M50) || defined(CONFIG_SX70) || defined(CONFIG_SX740) || defined(CONFIG_R) || defined(CONFIG_RP) || defined (CONFIG_250D) || defined(CONFIG_XF605)
-// variants M50, SX740, R + RP, 250D combined
-struct prop_lens_static_data
-{
-        uint8_t                 attached;
-        uint8_t                 type;                         // 90 EF, 91 RF
-        uint8_t                _unk_01[38];                   // Not referenced in readid
-        uint16_t                lens_id;
-        uint16_t                lens_id_ext;
-        uint16_t                fl_wide;
-        uint16_t                fl_tele;
-        uint8_t                 lens_serial[5];
-        uint8_t                _unk_02[24];                    // Not referenced in readid
-        uint8_t                 extender_id[6];
-        uint8_t                 lens_firm_ver[3];
-        uint8_t                 field_vision;
-        uint8_t                 lens_type;
-#if defined(CONFIG_R) || defined(CONFIG_RP)
-        uint8_t                _pad_01;                        // padding exists on R,RP
-#endif // defined(CONFIG_R) || defined(CONFIG_RP)
-        uint8_t                 lens_name_len;
-        char                    lens_name[73];
-        uint8_t                _unk_03;                        // Not referenced in readid
-        uint8_t                 mount_size;
-        uint8_t                 lens_switch_exists;
-        uint8_t                 lens_is_switch_exists;
-        uint8_t                 lens_is_funct_exists;
-        uint8_t                 af_speed_setting_available;
-#if !defined(CONFIG_M50)
-        uint8_t                 dafLimitFno;
-        uint8_t                 distortionCorrectionInfo;
-        uint8_t                 bcfInfo;
-        uint8_t                _unk_04;                        // Not referenced in readid
-#if !defined(CONFIG_250D) && !defined(CONFIG_SX70)
-        uint8_t                _pad_02;                        // 250D, SX70
-#endif // !defined(CONFIG_250D) && !defined(CONFIG_SX70)
-#endif // !defined(CONFIG_M50)
-        uint16_t                zoom_pos_size;
-        uint16_t                focus_pos_size;
-        uint16_t                fine_focus_size;
-#if !defined(CONFIG_M50)
-        uint8_t                 av_dlp_lens;
-        uint8_t                 av_slow_enable;
-#endif // !defined(CONFIG_M50)
-        uint8_t                 av_slow_div;
-        uint8_t                _unk_05;                        // Not referenced in readid
-        uint16_t                av_max_spd;
-        uint16_t                av_silent_spd;
-        uint16_t                av_min_spd;
-#if defined(CONFIG_M50)
-        uint8_t                _unk_06[95];                    // Not referenced in readid
-#elif defined(CONFIG_250D) || defined(CONFIG_SX70)
-        uint8_t                _unk_06[149];                   // Not referenced in readid
-#else // R, RP, looks like additional padding vs 250D exists
-        uint8_t                _unk_06[151];                   // Not referenced in readid
-#endif
-        uint8_t                 colorBalance;
-        uint8_t                 pza_exists;
-        uint8_t                 pza_id[5];
-        uint8_t                 pza_firm_ver[3];
-        uint8_t                 pza_firmup;
-        uint8_t                 dlAdp_count;
-        uint8_t                _unk_07[3];                     // Not referenced in readid
-        uint32_t                dlAdpl_id;
-        uint8_t                 dlAdpl_funcl;
-        uint8_t                 dlAdpl_firm_ver[3];
-        uint32_t                dlAdpl2_id;
-        uint8_t                 dlAdpl2_funcl;
-        uint8_t                 dlAdpl2_firm_ver[3];
-        uint8_t                 safemode_info;                 // named only on M50
-        uint8_t                 demandWarnDispFromLens;        // unnamed on M50
-        uint8_t                 demandWarnDispFromAdp;         // unnamed on M50
-#if defined(CONFIG_M50)
-        uint8_t                _unk_08;                        // Not referenced in readid
-#else
-        uint8_t                _unk_08[13];                    // Not referenced in readid
-#endif
-#if defined(CONFIG_XF605)
-// FIXME struct is unknown, all preceding fields may be junk.
-// Size is known however.  See prop_request_change() usage
-// for PROP_LENS_STATIC_DATA.
-        uint8_t                _unk_09[0x14c];
-#endif
-};
-
-    #if defined(CONFIG_SX740)
-        #pragma message "FIXME: SX740 prop_lens_static_data is not implemented"
-    #endif
-    #if defined(CONFIG_M50)
-        SIZE_CHECK_STRUCT( prop_lens_static_data, 0x138 );
-        //#elif defined(CONFIG_SX740) kitor FIXME: enable
-        //SIZE_CHECK_STRUCT( prop_lens_static_data, 0x178 );
-    #elif defined(CONFIG_250D) || defined(CONFIG_SX70)
-        SIZE_CHECK_STRUCT( prop_lens_static_data, 0x180 );
-    #elif defined(CONFIG_XF605)
-        SIZE_CHECK_STRUCT( prop_lens_static_data, 0x2d0 );
-    #else  // R, RP, SX70
-        SIZE_CHECK_STRUCT( prop_lens_static_data, 0x184 );
-    #endif // size check M50, R, RP, 250D
-
-#elif defined(CONFIG_850D) || defined(CONFIG_R6) || defined(CONFIG_R5)
-/* new struct variant reorders some fields as compared to previous
- * thus making a separate definition */
-struct prop_lens_static_data
-{
-        uint8_t                 attached;
-        uint8_t                 type;
-        uint8_t                _unk_01[38];                    // Not referenced in readid
-        uint16_t                lens_id;
-        uint16_t                lens_id_ext;
-        uint16_t                fl_wide;
-        uint16_t                fl_tele;
-        uint8_t                 lens_serial[5];
-        uint8_t                _unk_02[24];                    // Not referenced in readid
-        uint8_t                 extender_id[6];
-        uint8_t                 lens_firm_ver[3];
-        uint8_t                 field_vision;
-        uint8_t                 lens_type;
-        uint8_t                 extenderMountInfo;
-        uint8_t                 lens_name_len;
-        char                    lens_name[73];
-        uint8_t                _unk_03[65];                    // Not referenced in readid
-        uint8_t                 mount_size;
-        uint8_t                 lens_switch_exists;
-        uint8_t                 lens_af_func_exists;
-        uint8_t                 lens_mf_funct_exists;
-        uint8_t                 lens_is_switch_exists;
-        uint8_t                 lens_is_funct_exists;
-        uint8_t                 af_speed_setting_available;
-        uint8_t                 dafLimitFno;
-        uint8_t                 distortionCorrectionInfo;
-        uint8_t                 bcfInfo;
-        uint8_t                 lens_id_1292;
-        uint8_t                 emd_hot_limit;
-#if defined(CONFIG_R6) || defined(CONFIG_R5)
-        uint8_t                 aberationControl;              // DNE on 850D
-        uint8_t                 _pad_01;
-#endif // defined(CONFIG_R6) || defined(CONFIG_R5)
-        uint16_t                zoom_pos_size;
-        uint16_t                focus_pos_size;
-        uint16_t                fine_focus_size;
-        uint16_t                av_slow_div;
-        uint8_t                 av_slow_enable;
-        uint8_t                 av_dlp_lens;
-        uint16_t                av_dlp_max_spd;
-        uint16_t                av_dlp_silent_spd;
-        uint16_t                av_dlp_min_spd;
-        uint8_t                _unk_04[150];
-        uint8_t                 extendMagnificationVal;
-        uint8_t                _unk_05;
-        uint16_t                ois_shift_max;
-#if defined(CONFIG_R6) || defined(CONFIG_R5)
-        uint8_t                 colorBalance;                  // DNE on 850D
-#endif // defined(CONFIG_R6) || defined(CONFIG_R5)
-        uint8_t                 pza_exists;
-        uint8_t                 pza_id[5];
-        uint8_t                 pza_firm_ver[3];
-        uint8_t                 pza_firmup;
-        uint8_t                 dlAdp_count;
-#if defined(CONFIG_850D)
-        uint8_t                _pad_02[3];
-#endif //defined(CONFIG_850D)
-        uint32_t                dlAdpl_id;
-        uint8_t                 dlAdpl_funcl;
-        uint8_t                 dlAdpl_firm_ver[3];
-        uint32_t                dlAdpl2_id;
-        uint8_t                 dlAdpl2_funcl;
-        uint8_t                 dlAdpl2_firm_ver[3];
-        uint8_t                _unk_06;
-        uint8_t                 demandWarnDispFromLens;
-        uint8_t                 demandWarnDispFromAdp;
-        uint8_t                _unk_07[13];
-};
-
-SIZE_CHECK_STRUCT( prop_lens_static_data, 0x1C8 );
-#else  // unknown model
-#error No PROP_LENS_STATIC_DATA defined for built cam model
-#endif // unknown model
-
-/*
- * PROP_LENS_DYNAMIC_DATA structure
- *
- * For PROP_LENS_DYNAMIC_DATA it is a bit hard to track in decomp. Full content
- * is copied a lot between different subsystems.
- *
- * However after a lot of tracking I found function that names a lot of fields.
- * See _prints_dynamic_data(dynamic*, static*).
- * M50.110: e06e6dbc, R180: e02307b8, 250D.100: e06469f8...
- * Function easy to trace with print starting with "AVEF:0x%x"(...)
- *
- * There's also another function, see R180 e0381a64. On param2 it takes
- * "ShootingInfoEx" structure. This struct is a bundle of multiple structs,
- * including lens dynamic data.
- * Scroll to a debug print "ShootingInfoEx:LensDynamicInfo avef:%x"(...)
- * First field of "ShootingInfoEx" used by that print is also first field of
- * prop_lens_dynamic_data embedded into "ShootingInfoEx".
- */
-struct prop_lens_dynamic_data {
-        uint16_t                AVEF;             // ShootingInfoEx: avef
-        uint16_t                AVO;              // ShootingInfoEx: avo
-        uint16_t                AVMAX;            // ShootingInfoEx: avmax
-#if !defined(CONFIG_M50)
-        uint16_t                AVD;              // Not referenced in M50
-#if defined(CONFIG_850D) || defined(CONFIG_R6)  || defined(CONFIG_R5)
-        uint16_t                NowAvRF;          // Referenced 850D, R6
-#endif
-        uint16_t                NowAvEF;          // Not referenced in M50. Before 850D named just NowAv
-#endif
-        uint8_t                _pad_01[22];       // M50, R, RP, 250D, 850D, R6
-        uint16_t                jsstep;
-        uint8_t                _pad_02[4];        // M50, R, RP, 250D, 850D, R6
-        uint16_t                IDC;
-#if !defined(CONFIG_M50) && !defined(CONFIG_250D) && !defined(CONFIG_SX740)
-        uint8_t                _pad_03[2];        // R, RP, 850D, R6; not on M50, 250D (alignment?)
-#endif
-        uint16_t                po;
-        uint16_t                po05;
-        uint16_t                po10;
-        uint16_t                po00;
-        uint16_t                po15;
-        uint16_t                Npo;
-        uint16_t                po0;
-        uint16_t                po25;
-        uint16_t                po50;
-        uint16_t                po75;
-        uint16_t                po100;            // po100+ not referenced on 850D
-        uint16_t                po125;
-        uint16_t                po150;
-        uint16_t                po175;
-        uint16_t                po200;
-        uint8_t                 v0;               // vignetting
-        uint8_t                 v1;
-        uint8_t                 v2;
-        uint8_t                 v3;
-        uint16_t                FL;
-        uint16_t                focal_len_image;  // ShootingInfoEx R6, R
-        uint16_t                focus_far;        // ShootingInfoEx: abs_inf; FarAbs
-        uint16_t                focus_near;       // ShootingInfoEx: abs_near; NearAbs
-        uint16_t                zoomPos;          // ShootingInfoEx: zoom_pos
-        uint16_t                focusPos;         //
-        uint16_t                fineFocusPos;     // ShootingInfoEx: fine_focus_pos
-        uint16_t                HighResoZoomPos;  // ShootingInfoEx: high_res_zoom_pos
-        uint16_t                HighResoFocusPos; // ShootingInfoEx: high_res_focus_pos
-#if defined(CONFIG_R5) || defined(CONFIG_R6)
-        uint8_t                _r6_01[6];         // only on R6, some extra fields?
-#endif
-#if defined(CONFIG_SX70) || defined(CONFIG_R) || defined(CONFIG_RP) || defined(CONFIG_R5) || defined(CONFIG_R6)
-        uint8_t                 abstat;           // lens abberation related; exists only on R series
-#endif
-        uint8_t                 st1;
-        uint8_t                 st2;              // & 0x80 true -> MF, false -> AF
-        uint8_t                 st3;              // & 0x0F looks equv to PROP_LV_LENS_STABILIZE
-        uint8_t                 st4;
-        uint8_t                 st5;
-        uint8_t                _pad_04[3];        // M50, R, RP, 250D, 850D, R6
-        uint8_t                 FcsSt1;           // FocusStep?
-        uint8_t                 FcsSt2;           // FcsSt* are unnamed on M50
-        uint8_t                 FcsSt3;
-        uint8_t                 FcsSt4;
-        uint8_t                 ZmSt1;            // ZoomStep?
-        uint8_t                 ZmSt2;            // ZmSt* are unnamed on M50
-        uint8_t                 ZmSt3;
-        uint8_t                 ZmSt4;
-        uint8_t                _pad_05[4];        // M50, R, RP, 250D, 850D, R6
-#if defined(CONFIG_SX70) || defined(CONFIG_R) || defined(CONFIG_RP) || defined(CONFIG_R5) || defined(CONFIG_R6)
-        uint8_t                _pad_05a;          // R, RP, R6 (alignment?)
-#endif
-        uint16_t                ts_shift;         // via ShootingInfoEx
-        uint8_t                 ts_tilt;          // via ShootingInfoEx
-        uint8_t                 ts_all_revo;      // via ShootingInfoEx
-        uint8_t                 ts_ts_revo;       // via ShootingInfoEx
-        uint8_t                _pad_06[7];        // M50, R, RP, 250D, 850D, R6
-#if defined(CONFIG_M50)
-        uint8_t                _m50_01[4];        // M50, some extra fields?
-#endif
-        uint8_t                 LENSEr;           // not mentioned on 850D
-#if defined(CONFIG_M50)
-        uint8_t                _pad_07[7];        // M50
-#elif defined(CONFIG_R5) || defined(CONFIG_R6)
-        uint8_t                _pad_07[11];       // R6
-#else
-        uint8_t                _pad_07[15];       // 850D, 250D, R, RP
-#endif
-#if defined(CONFIG_XF605)
-        uint8_t                _pad_08[0x136]; // 0x1c4 total
-#endif
-};
-
-#if defined(CONFIG_R5) || defined(CONFIG_R6)
-SIZE_CHECK_STRUCT( prop_lens_dynamic_data, 0x94 );
-#elif defined(CONFIG_SX70) || defined(CONFIG_R) || defined(CONFIG_RP) || defined(CONFIG_850D)
-SIZE_CHECK_STRUCT( prop_lens_dynamic_data, 0x90);
-#elif defined(CONFIG_250D) || defined(CONFIG_SX740)
-SIZE_CHECK_STRUCT( prop_lens_dynamic_data, 0x8C);
-#elif defined(CONFIG_M50)
-SIZE_CHECK_STRUCT( prop_lens_dynamic_data, 0x84);
-#elif defined(CONFIG_XF605)
-// FIXME the dynamic and static structs for XF605 are almost certainly very wrong.
-// Couldn't find a nice debug function that dumped the names of the fields.
-// This is just to let a build compile and shouldn't be used.  The sizes are
-// believed to be correct.
-SIZE_CHECK_STRUCT( prop_lens_dynamic_data, 0x1c4); // see e.g. 1ed6ecae on XF605 1.0.1
-#else  // unknown model
-#error No PROP_LENS_DYNAMIC_DATA defined for built cam model
-#endif // /unknown model
-
-#endif // CONFIG_DIGIC_VIII + CONFIG_DIGIC_X
 
 struct prop_focus
 {
